Beatified scripts

+++ Assets/Scripts/SaveSystem.cs
@@ -5,7 +5,6 @@ using System.Runtime.Serialization.Formatters.Binary;
 
 public static class SaveSystem 
 {
-    
     public static readonly string PlayerPath = Application.persistentDataPath + "PlayerSave.data";
     public static readonly string WorldPath = Application.persistentDataPath + "WorldSave.data";
 
@@ -13,11 +12,9 @@ public static class SaveSystem
     {
         BinaryFormatter bf = new BinaryFormatter();
         FileStream stream = new FileStream(PlayerPath, FileMode.Create);
-
         PlayerData data = new PlayerData(player);
         
         bf.Serialize(stream, data);
-        
         stream.Close();
     }
     
@@ -36,27 +33,27 @@ public static class SaveSystem
             Debug.LogError("World save file not found in " + WorldPath);
             throw;
         }
+        
         bf.Serialize(stream, data);
         stream.Close();
     }
 
     public static PlayerData LoadPlayerData()
+    {
+        if (File.Exists(PlayerPath))
         {
-            if (File.Exists(PlayerPath))
-            {
-                BinaryFormatter bf = new BinaryFormatter(); 
-                FileStream stream = new FileStream(PlayerPath, FileMode.Open);
-                
-                PlayerData data = bf.Deserialize(stream) as PlayerData;
-                
-                stream.Close();
-                return data;
-            }
+            BinaryFormatter bf = new BinaryFormatter(); 
+            FileStream stream = new FileStream(PlayerPath, FileMode.Open);
             
-            Debug.LogError("Player save file not found in " + PlayerPath);
-            return null; 
+            PlayerData data = bf.Deserialize(stream) as PlayerData;
+            stream.Close();
             
+            return data;
         }
+        
+        Debug.LogError("Player save file not found in " + PlayerPath);
+        return null;
+    }
 
     public static WorldData LoadWorldData()
     {
@@ -66,21 +63,18 @@ public static class SaveSystem
             FileStream stream = new FileStream(WorldPath, FileMode.Open);
             
             WorldData data = bf.Deserialize(stream) as WorldData;
-            
             stream.Close();
+            
             return data;
         }
         
         Debug.LogError("World save file not found in " + WorldPath);
         return null;
-        
     }
     
     public static void DeleteData()
     {
-        if (File.Exists(PlayerPath))
-            File.Delete(PlayerPath);
-        if (File.Exists(WorldPath))
-            File.Delete(WorldPath);
+        if (File.Exists(PlayerPath)) File.Delete(PlayerPath);
+        if (File.Exists(WorldPath)) File.Delete(WorldPath);
     }
 }
